+++ truck-base/src/newton.rs
@@ -1,73 +1,84 @@
 //! Implementation of Newton method
-//!
-//! # Examples
-//! ```
-//! use truck_base::{newton::newton1, assert_near2};
-//! 
-//! let value = |x: f64| x * x - 2.0;
-//! let der = |x: f64| 2.0 * x;
-//! let sqrt2 = newton1(value, der, 1.0, 10).unwrap();
-//! assert_near2!(sqrt2, f64::sqrt(2.0));
-//! ```
+
+use std::ops::{Mul, Sub};
 
 use crate::{cgmath64::*, tolerance::*};
 
-macro_rules! define_newton {
-    ($method_name: ident, $matrix: ty, $vector: ty, ($($ders: ident),*), $doc: expr) => {
-        #[doc = $doc]
-        pub fn $method_name(
-            value: impl Fn($vector) -> $vector,
-            $($ders: impl Fn($vector) -> $vector,)*
-            mut hint: $vector,
-            trials: usize,
-        ) -> Result<$vector, NewtonLog<$vector>> {
-            let mut log = NewtonLog::new(cfg!(debug_assertions), trials);
-            for _ in 0..=trials {
-                log.push(hint);
-                let mat = <$matrix>::from_cols($($ders (hint),)*);
-                let Some(inv) = mat.invert() else {
-                    log.set_degenerate(true);
-                    return Err(log);
-                };
-                let next = hint - inv * value(hint);
-                if next.near2(&hint) {
-                    return Ok(next);
-                }
-                hint = next;
-            }
-            Err(log)
+/// the value and jacobian corresponding to parameter
+#[derive(Clone, Debug)]
+pub struct CalcOutput<V, M> {
+    /// the value of function
+    pub value: V,
+    /// the jacobian of function
+    pub derivation: M,
+}
+
+/// jacobian of function
+pub trait Jacobian<V>: Mul<V, Output = V> + Sized {
+    #[doc(hidden)]
+    fn invert(self) -> Option<Self>;
+}
+
+impl Jacobian<f64> for f64 {
+    #[inline(always)]
+    fn invert(self) -> Option<Self> {
+        match self.is_zero() {
+            true => None,
+            false => Some(1.0 / self),
+        }
+    }
+}
+
+macro_rules! impl_jacobian {
+    ($matrix: ty, $vector: ty) => {
+        impl Jacobian<$vector> for $matrix {
+            #[inline(always)]
+            fn invert(self) -> Option<Self> { SquareMatrix::invert(&self) }
         }
     };
 }
 
-define_newton!(
-    newton1,
-    f64,
-    f64,
-    (der),
-    "Newton method in 1-dimension."
-);
-define_newton!(
-    newton2,
-    Matrix2,
-    Vector2,
-    (der0, der1),
-    "Newton method in 2-dimension."
-);
-define_newton!(
-    newton3,
-    Matrix3,
-    Vector3,
-    (der0, der1, der2),
-    "Newton method in 3-dimension."
-);
-define_newton!(
-    newton4,
-    Matrix4,
-    Vector4,
-    (der0, der1, der2, der3),
-    "Newton method in 4-dimension."
-);
+impl_jacobian!(Matrix2, Vector2);
+impl_jacobian!(Matrix3, Vector3);
+impl_jacobian!(Matrix4, Vector4);
+
+/// Solve equation by Newton's method
+/// # Examples
+/// ```
+/// use truck_base::{newton::*, assert_near2};
+///
+/// let function = |x: f64| CalcOutput {
+///     value: x * x - 2.0,
+///     derivation: 2.0 * x,
+/// };
+/// let sqrt2 = solve(function, 1.0, 10).unwrap();
+/// assert_near2!(sqrt2, f64::sqrt(2.0));
+/// ```
+pub fn solve<V, M>(
+    function: impl Fn(V) -> CalcOutput<V, M>,
+    mut hint: V,
+    trials: usize,
+) -> Result<V, NewtonLog<V>>
+where
+    V: Sub<Output = V> + Copy + Tolerance,
+    M: Jacobian<V>,
+{
+    let mut log = NewtonLog::new(cfg!(debug_assertions), trials);
+    for _ in 0..=trials {
+        log.push(hint);
+        let CalcOutput { value, derivation } = function(hint);
+        let Some(inv) = derivation.invert() else {
+            log.set_degenerate(true);
+            return Err(log);
+        };
+        let next = hint - inv * value;
+        if next.near2(&hint) {
+            return Ok(next);
+        }
+        hint = next;
+    }
+    Err(log)
+}
 
 mod newtonlog {
     use std::fmt::*;
@@ -97,7 +108,11 @@ mod newtonlog {
         #[inline(always)]
         pub fn degenerate(&self) -> bool { self.degenerate }
         #[inline(always)]
-        pub(super) fn push(&mut self, log: T) { self.log.as_mut().map(|vec| vec.push(log)); }
+        pub(super) fn push(&mut self, log: T) {
+            if let Some(vec) = &mut self.log {
+                vec.push(log)
+            }
+        }
         #[inline(always)]
         pub(super) fn set_degenerate(&mut self, degenerate: bool) { self.degenerate = degenerate }
     }
@@ -122,18 +137,3 @@ mod newtonlog {
     }
 }
 pub use newtonlog::NewtonLog;
-
-trait InvertFloat: Sized {
-    fn from_cols(self) -> Self { self }
-    fn invert(self) -> Option<Self>;
-}
-
-impl InvertFloat for f64 {
-    #[inline(always)]
-    fn invert(self) -> Option<Self> {
-        match self.is_zero() {
-            true => None,
-            false => Some(1.0 / self),
-        }
-    }
-}