+++ truck-geometry/src/nurbs/nurbssurface.rs
@@ -527,7 +527,7 @@ impl<V: Homogeneous<f64>> SearchNearestParameter<D2> for NurbsSurface<V>
 where
     Self: ParametricSurface<Point = V::Point, Vector = <V::Point as EuclideanSpace>::Diff>,
     V::Point: EuclideanSpace<Scalar = f64> + MetricSpace<Metric = f64>,
-    <V::Point as EuclideanSpace>::Diff: SspVector,
+    <V::Point as EuclideanSpace>::Diff: SspVector<Point = V::Point>,
 {
     type Point = V::Point;
     /// Searches the parameter `(u, v)` which minimize `|self(u, v) - point|` by Newton's method
@@ -765,7 +765,7 @@ where
         + EuclideanSpace<Scalar = f64, Diff = V>
         + MetricSpace<Metric = f64>
         + Tolerance,
-    V: SspVector,
+    V: SspVector<Point = P>,
 {
     type Point = P;
     /// Search the parameter `(u, v)` such that `self.subs(u, v).rational_projection()` is near `pt`.